+++ R/massdefect.R
@@ -25,9 +25,9 @@ getmassdefect <- function(mass, sf) {
 #' @param submass mass vector of sub structure of homologous series
 #' @param mdcutoff mass defect cluster cutoff
 #' @return list with tentative isotope, adducts, and neutral loss peaks' index, retention time cluster, std mass defect analysis dataframe
-#' @seealso \code{\link{getmassdefect}},\code{\link{plotkms}}
+#' @seealso \code{\link{getmassdefect}},\code{\link{plotkms}},\code{\link{getcorstd}}
 #' @export
-getpaired <- function(list, rtcutoff = 9, isocutoff = 3, freqcutoff = 20, submass = c(15.9949,14.01565,43.00581,30.01056,34.96885,78.91834), mdcutoff = 0.02){
+getpaired <- function(list, rtcutoff = 9, isocutoff = 3, freqcutoff = 20, submass = c(15.9949,14.003074,26.01568,14.01565,43.00581,30.01056,34.96885,78.91834), mdcutoff = 0.02){
 
         # paired mass diff analysis
         groups <- cbind.data.frame(mz = list$mz, rt = list$rt)
@@ -124,6 +124,70 @@ getpaired <- function(list, rtcutoff = 9, isocutoff = 3, freqcutoff = 20, submas
 
         list$rtcluster <- rtcluster
 
+        list$stdmassindex <- (round(list$mz,4) %in% round(resultstd$mz,4))
+        list$stdmass <- resultstd
+        return(list)
+}
+
+#' Paired correlationship among peak list based on cluster analysis
+#' @param list a list with mzrt profile
+#' @param rtcutoff cutoff of the distances in cluster
+#' @param submass mass vector of sub structure of homologous series
+#' @param mdcutoff mass defect cluster cutoff
+#' @return list with retention time cluster, std mass defect analysis dataframe based on max average correlation
+#' @seealso \code{\link{getmassdefect}},\code{\link{plotkms}},\code{\link{getpaired}}
+#' @export
+getcorstd <- function(list, rtcutoff = 9, submass = c(15.9949,14.003074,26.01568,14.01565,43.00581,30.01056,34.96885,78.91834), mdcutoff = 0.02){
+
+        # paired mass diff analysis
+        groups <- cbind.data.frame(mz = list$mz, rt = list$rt, list$data)
+        resultstd <- resultsolo <- resultiso <- result <- NULL
+
+        dis <- stats::dist(list$rt, method = "manhattan")
+        fit <- stats::hclust(dis)
+        rtcluster <- stats::cutree(fit, h=rtcutoff)
+        n <- length(unique(rtcluster))
+        message(paste(n, 'retention time cluster found.'))
+        # search:
+        for (i in 1:length(unique(rtcluster))) {
+                # find the mass within RT
+                rtxi <- list$rt[rtcluster == i]
+                bin = groups[groups$rt %in% rtxi, ]
+                medianrtxi <- stats::median(rtxi)
+
+                if (nrow(bin) > 1) {
+                        # get mz diff
+                        cor <- stats::cor(t(bin[,-c(1,2)]))
+                        cormean <- apply(cor,1,mean)
+                        corindex <- which.max(cormean)
+                        df <- cbind(bin[corindex,],rtg = i)
+                        result <- rbind(result,df)
+                }else{
+                        solo <- cbind(bin,rtg = i)
+                        resultsolo <- rbind(solo,resultsolo)
+                }
+        }
+
+        resultstd <- rbind(result,resultsolo)
+        resultstd <- unique(resultstd)
+
+        # perform mass defect analysis for std mass
+        for(i in 1:length(submass)){
+                mdst <- round(submass[i])/submass[i]
+                msdefect <- round(resultstd$mz*mdst) - resultstd$mz*mdst
+                dis <- stats::dist(msdefect, method = "manhattan")
+                fit <- stats::hclust(dis)
+                mdcluster <- stats::cutree(fit, h=mdcutoff)
+                n <- length(unique(mdcluster))
+                message(paste(n, 'mass defect clusters found for mass', submass[i], 'substructures' ))
+                name <- c(colnames(resultstd),submass[i],paste0(submass[i],'g'))
+                resultstd <- cbind.data.frame(resultstd,msdefect,mdcluster)
+                colnames(resultstd) <- name
+        }
+
+
+        # filter the list
+        list$rtcluster <- rtcluster
         list$stdmassindex <- (round(list$mz,4) %in% round(resultstd$mz,4))
         list$stdmass <- resultstd
         return(list)